feat: implement BlockService to manage block creation and hash calculation
