+++ wallet/src/main/kotlin/net/thechance/wallet/service/PaymentService.kt
@@ -1,17 +1,12 @@
 package net.thechance.wallet.service
 
-import net.thechance.wallet.entity.Block
 import net.thechance.wallet.entity.PendingTransaction
 import net.thechance.wallet.entity.toTransaction
-import net.thechance.wallet.repository.BlockRepository
 import net.thechance.wallet.repository.PendingTransactionRepository
 import net.thechance.wallet.repository.TransactionRepository
-import org.springframework.data.domain.Pageable
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.security.MessageDigest
-import java.time.LocalDateTime
 import java.util.*
 
 @Service
@@ -19,14 +14,14 @@ class PaymentService(
     private val pendingTransactionRepository: PendingTransactionRepository,
     private val transactionRepository: TransactionRepository,
     private val walletService: WalletService,
-    private val blockRepository: BlockRepository
+    private val blockService: BlockService
 ) {
 
     @Transactional
     fun pay(userId: UUID, transactionId: UUID) {
         val pendingTransaction = validateAndGetTransaction(userId, transactionId)
 
-        val block = getCurrentBlock()
+        val block = blockService.getCurrentBlock()
         val transaction = pendingTransaction.toTransaction(block)
         transactionRepository.save(transaction)
         pendingTransactionRepository.deleteById(transactionId)
@@ -49,41 +44,5 @@ class PaymentService(
             throw IllegalArgumentException("Insufficient balance")
 
         return pendingTransaction
-    }
-
-    private fun getCurrentBlock(): Block {
-        val latestBlock = blockRepository.findTopByOrderByTimestampDesc()
-        if (latestBlock != null) {
-            val transactionCount = transactionRepository.countAllByBlockId(latestBlock.id)
-            if (transactionCount < TRANSACTION_COUNT_LIMIT_PER_BLOCK) {
-                return latestBlock
-            }
-        }
-        val previousBlockHash = calculatePreviousBlockHash(latestBlock)
-        return blockRepository.save(Block(previousBlockHash = previousBlockHash))
-    }
-
-    private fun calculatePreviousBlockHash(latestBlock: Block?): String {
-        if (latestBlock == null) return "0".repeat(64)
-
-        val transactionsData = transactionRepository
-            .getAllByBlockId(latestBlock.id, Pageable.ofSize(TRANSACTION_COUNT_LIMIT_PER_BLOCK))
-            .joinToString(separator = "|") { it.toString() }
-
-        val input = latestBlock.id.toString() +
-                latestBlock.timestamp.toString() +
-                transactionsData +
-                LocalDateTime.now().toString()
-
-        return hashWithSha256(input)
-    }
-
-    private fun hashWithSha256(input: String): String {
-        val bytes = MessageDigest.getInstance("SHA-256").digest(input.toByteArray())
-        return bytes.joinToString("") { "%02x".format(it) }
-    }
-
-    private companion object {
-        const val TRANSACTION_COUNT_LIMIT_PER_BLOCK = 15
     }
 }
